feature(OneSignal): Adding comments
